Bump up the z-index of docs menu on small screens
On some docs pages, we have absolutely positioned content. Because the navigation comes *before* the content, that means content shows up in front of the menu. This gives the menu an explicit z-index so that it always sits in front of the content (unless content explicitly sets an even higher z-index, of course).