Correct some examples using docrun
Ran docrun over just two of our markdown files to see how it did. Fixed a few problems, but also added a bunch of TODOs to come back to later, once docrun has gained some more features. This gives some good indication of how much of qri we need to be able to mock-out in order to run all of these examples.